Moved `file.open` command to `@theia/core`; #1609
