+++ packages/git/src/browser/diff/git-diff-contribution.ts
@@ -12,7 +12,7 @@ import { injectable, inject } from "inversify";
 import { GitDiffWidget, GIT_DIFF } from './git-diff-widget';
 import { open, OpenerService } from "@theia/core/lib/browser";
 import { NAVIGATOR_CONTEXT_MENU } from '@theia/navigator/lib/browser/navigator-contribution';
-import { UriCommandHandler, UriAwareCommandHandler } from '@theia/workspace/lib/browser/workspace-commands';
+import { UriCommandHandler, UriAwareCommandHandler } from '@theia/core/lib/common/uri-command-handler';
 import { GitQuickOpenService } from '../git-quick-open-service';
 import { FileSystem } from "@theia/filesystem/lib/common";
 import { DiffUris } from '@theia/core/lib/browser/diff-uris';
